+++ gui-builder/include/WidgetProperties/SpinControlProperties.hpp
@@ -43,8 +43,8 @@ struct SpinControlProperties : WidgetProperties
             spinControl->setStep(value.toFloat());
         else if (property == "DecimalPlaces")
             spinControl->setDecimalPlaces(value.toUInt());
-        else if (property == "UseWideArrows")
-            spinControl->setUseWideArrows(parseBoolean(value, false));
+        else if (property == "SpinButtonWidth")
+            spinControl->setSpinButtonWidth(value.toFloat());
         else
             WidgetProperties::updateProperty(widget, property, value);
     }
@@ -57,9 +57,8 @@ struct SpinControlProperties : WidgetProperties
         pair.first["Maximum"] = {"Float", tgui::String::fromNumber(spinControl->getMaximum())};
         pair.first["Value"] = {"Float", tgui::String::fromNumber(spinControl->getValue())};
         pair.first["Step"] = {"Float", tgui::String::fromNumber(spinControl->getStep())};
-        pair.first["Step"] = {"Float", tgui::String::fromNumber(spinControl->getStep())};
         pair.first["DecimalPlaces"] = {"UInt", tgui::String::fromNumber(spinControl->getDecimalPlaces())};
-        pair.first["UseWideArrows"] = {"Bool", tgui::Serializer::serialize(spinControl->getUseWideArrows())};
+        pair.first["SpinButtonWidth"] = {"Float", tgui::String::fromNumber(spinControl->getSpinButtonWidth())};
 
         const auto buttonRenderer = spinControl->getSpinButtonSharedRenderer();
         pair.second["SpinButton.ButtonsBorders"] = {"Outline", tgui::Serializer::serialize(buttonRenderer->getBorders())};
@@ -89,9 +88,6 @@ struct SpinControlProperties : WidgetProperties
         pair.second["SpinText.BackgroundColorFocused"] = {"Color", tgui::Serializer::serialize(textRenderer->getBackgroundColorFocused())};
         pair.second["SpinText.CaretColor"] = {"Color", tgui::Serializer::serialize(textRenderer->getCaretColor())};
         pair.second["SpinText.CaretColorHover"] = {"Color", tgui::Serializer::serialize(textRenderer->getCaretColorHover())};
-TGUI_IGNORE_DEPRECATED_WARNINGS_START
-        pair.second["SpinText.CaretColorFocused"] = {"Color", tgui::Serializer::serialize(textRenderer->getCaretColorFocused())};
-TGUI_IGNORE_DEPRECATED_WARNINGS_END
         pair.second["SpinText.BorderColor"] = {"Color", tgui::Serializer::serialize(textRenderer->getBorderColor())};
         pair.second["SpinText.BorderColorHover"] = {"Color", tgui::Serializer::serialize(textRenderer->getBorderColorHover())};
         pair.second["SpinText.BorderColorDisabled"] = {"Color", tgui::Serializer::serialize(textRenderer->getBorderColorDisabled())};